[chore] Split keccak implementation into multiple files (#120)
Split keccak implementation into multiple files
